Extract config to separate files
Why:

* Gatsby uses a small number of files in the root of the project to
  configure the website, which we'll need to use to generate the blog
  posts pages. The issue here is that these configs make these files too
  big too fast, and at some point I was starting to feel some pain
  navigating this code.

This change addresses the issue by:

* Using a pattern I've seen but don't remember where, which separated
  each Gatsby API _hook_ into its own file. This works very well since
  each such _hook_ is, in some way, self-contained. So, we can isolate
  the logic required for each hook in a single file, keeping the files
  in the root a lot simpler.
