+++ config/site_metadata.js
@@ -1,0 +1,7 @@
+module.exports = {
+  title: "Subvisual",
+  description: "We nurture ıdeas that empower people",
+  url: "https://subvisual.com",
+  image: "/images/meta-image.jpg",
+  twitterUsername: "@subvisual",
+}
