+++ gatsby-config.js
@@ -1,77 +1,4 @@
-const postCssCustomMedia = require("postcss-custom-media")
+const plugins = require("./config/plugins")(__dirname)
+const siteMetadata = require("./config/site_metadata")
 
-module.exports = {
-  siteMetadata: {
-    title: "Subvisual",
-    description: "We nurture ıdeas that empower people",
-    url: "https://subvisual.com",
-    image: "/images/meta-image.jpg",
-    twitterUsername: "@subvisual",
-  },
-  plugins: [
-    {
-      resolve: "gatsby-plugin-sass",
-      options: {
-        postCssPlugins: [postCssCustomMedia()],
-      },
-    },
-    "gatsby-plugin-react-helmet",
-    "gatsby-transformer-yaml",
-    {
-      resolve: "gatsby-source-filesystem",
-      options: {
-        path: `${__dirname}/src/data`,
-      },
-    },
-    {
-      resolve: "gatsby-source-filesystem",
-      options: {
-        path: `${__dirname}/src/documents`,
-      },
-    },
-    {
-      resolve: "gatsby-source-filesystem",
-      options: {
-        name: "images",
-        path: `${__dirname}/src/images`,
-      },
-    },
-    "gatsby-transformer-sharp",
-    "gatsby-plugin-sharp",
-    {
-      resolve: "gatsby-plugin-manifest",
-      options: {
-        name: "subvisual",
-        short_name: "subvisual",
-        start_url: "/",
-        background_color: "#ffffff",
-        theme_color: "#ffffff",
-        display: "minimal-ui",
-        icon: "src/images/subvisual-symbol-blue.svg",
-      },
-    },
-    {
-      resolve: "gatsby-plugin-react-svg",
-      options: {
-        rule: {
-          include: /\.inline\.svg$/,
-        },
-      },
-    },
-    {
-      resolve: "gatsby-plugin-google-analytics",
-      options: {
-        trackingId: "UA-63400449-1",
-        head: false,
-        anonymize: true,
-        respectDNT: true,
-      },
-    },
-    {
-      resolve: "gatsby-plugin-netlify-cache",
-      options: {
-        cachePublic: true,
-      },
-    },
-  ],
-}
+module.exports = { plugins, siteMetadata }
